requirements.txt
pandas
numpy
matplotlib
seaborn
scikit-learn
mlflow
pymongo
from_root
certifi
imblearn
python-dotenv
ipykernel
pymongo[srv]==3.11
dill
pyaml

##-e .